% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{autoboost.res.x.adapt}
\alias{autoboost.res.x.adapt}
\title{Autoboost adaptative lasso diabetes first order.}
\format{A numerical matrix frame with 13 rows and 10 variables with attributes.}
\usage{
autoboost.res.x.adapt
}
\description{
Result of autoboost analysis of diabetes data from \cite{lars} package with adaptative
lasso and first order model
}
\keyword{datasets}
