% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{autoboost.res.x2.adapt}
\alias{autoboost.res.x2.adapt}
\title{Autoboost adaptative lasso diabetes second order.}
\format{
A numerical matrix frame with 13 rows and 64 variables with attributes.
}
\usage{
autoboost.res.x2.adapt
}
\description{
Result of autoboost analysis of diabetes data from \cite{lars} package with adaptative
lasso and second order model
}
\keyword{datasets}
