% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fastboost.res.x.adapt}
\alias{fastboost.res.x.adapt}
\title{Fastboost adaptative lasso diabetes first order.}
\format{A numerical matrix frame with 13 rows and 10 variables with attributes.}
\usage{
fastboost.res.x.adapt
}
\description{
Result of fastboost analysis of diabetes data from \cite{lars} package with adaptative
lasso and first order model
}
\keyword{datasets}
