% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fastboost.res.x2.adapt}
\alias{fastboost.res.x2.adapt}
\title{Fastboost adaptative lasso diabetes second order.}
\format{
A numerical matrix frame with 13 rows and 64 variables with attributes.
}
\usage{
fastboost.res.x2.adapt
}
\description{
Result of fastboost analysis of diabetes data from \cite{lars} package with adaptative
lasso and second order model
}
\keyword{datasets}
