% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscplots.R
\name{miscplot}
\alias{miscplot}
\alias{plot.matrix}
\title{Miscellaneous plot functions}
\usage{
\method{plot}{matrix}(x, ...)
}
\arguments{
\item{x}{A numeric matrix. A matrix to be plotted.}

\item{...}{. Additionnal arguments passed to the plot function.}
}
\value{
\code{matrixplot} returns \code{1}.
}
\description{
Define some additional plot functions to be used in the demos of the package.
}
\details{
\code{matrixplot} plots a numeric matrix \code{x}.
}
\examples{
set.seed(3141)
randmat=matrix(rnorm(360),60,60)
plot(randmat)

}
\references{
\emph{selectBoost: a general algorithm to enhance the performance of variable selection methods in correlated datasets}, Ismaïl Aouadi, Nicolas Jung, Raphael Carapito, Laurent Vallat, Seiamak Bahram, Myriam Maumy-Bertrand, Frédéric Bertrand, \url{https://arxiv.org/abs/1810.01670}
}
\author{
Frederic Bertrand, \email{frederic.bertrand@math.unistra.fr} with contributions from Nicolas Jung.
}
\concept{Plot functions}
