% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_confidence.R
\docType{class}
\name{network.confidence-class}
\alias{network.confidence-class}
\title{Network confidence class.}
\description{
Some details about this class and my plans for it in the body.
}
\details{
\describe{
   \item{network.confidence}{Matrix of confidence indices.}

   \item{name}{Vector.}

   \item{array}{F array}

   \item{time_pt}{Vector}

   \item{cv.subjects}{Logical. Was crossvalidation carried out subjectwise?}
 }
}
