% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoboost.R
\name{plot.summary.selectboost}
\alias{plot.summary.selectboost}
\title{Plot a summary of selectboost results}
\usage{
\method{plot}{summary.selectboost}(x, ...)
}
\arguments{
\item{x}{Numerical matrix. Summary of selectboost object.}

\item{...}{. Passed to the plotting functions.}
}
\value{
An invisible list.
}
\description{
Plot a summary of selectboost results.
}
\details{
\code{plot.summary.selectboost} returns an invisible list and creates four graphics.
Two plots the proportion of selection with respect to c0 (by step or according to real scale).
On the third graph, no bar means a proportion of selection less than prop.level.
Confidence intervals are computed at the conf.int.level level.
Barplot of the confidence index (1-min(c0, such that proportion|c0>conf.threshold)).
}
\examples{
set.seed(314)
xran=matrix(rnorm(75),15,5)
ybin=sample(0:1,15,replace=TRUE)
yran=rnorm(15)
layout(matrix(1:4,2,2))

data(autoboost.res.x)

plot.summary.selectboost(autoboost.res.x)

data(autoboost.res.x2)

plot.summary.selectboost(autoboost.res.x2)

}
\references{
\emph{selectBoost: a general algorithm to enhance the performance of variable selection methods in correlated datasets}, Ismaïl Aouadi, Nicolas Jung, Raphael Carapito, Laurent Vallat, Seiamak Bahram, Myriam Maumy-Bertrand, Frédéric Bertrand, \url{https://arxiv.org/abs/1810.01670}
}
\seealso{
\code{\link{fastboost}}, \code{\link{autoboost}} and \code{\link{summary.selectboost}}
}
\author{
Frederic Bertrand, \email{frederic.bertrand@math.unistra.fr}
}
\concept{Selectboost analyze functions}
