% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{results_simuls_reverse_engineering_v3}
\alias{results_simuls_reverse_engineering_v3}
\alias{test.seq_C}
\alias{test.seq_PL}
\alias{test.seq_PL2}
\alias{test.seq_PL2_W}
\alias{test.seq_PL2_tW}
\alias{test.seq_PSel}
\alias{test.seq_PSel.5}
\alias{test.seq_PSel.e2}
\alias{test.seq_PSel.5.e2}
\alias{test.seq_PSel_W}
\alias{test.seq_robust}
\alias{test.seq_PB}
\alias{test.seq_PB_095_075}
\alias{test.seq_PB_075_075}
\alias{test.seq_PB_W}
\alias{sensitivity_C}
\alias{sensitivity_PL}
\alias{sensitivity_PL2}
\alias{sensitivity_PL2_W}
\alias{sensitivity_PL2_tW}
\alias{sensitivity_PSel}
\alias{sensitivity_PSel.5}
\alias{sensitivity_PSel.e2}
\alias{sensitivity_PSel.5.e2}
\alias{sensitivity_PSel_W}
\alias{sensitivity_robust}
\alias{sensitivity_PB}
\alias{sensitivity_PB_095_075}
\alias{sensitivity_PB_075_075}
\alias{sensitivity_PB_W}
\alias{predictive_positive_value_C}
\alias{predictive_positive_value_PL}
\alias{predictive_positive_value_PL2}
\alias{predictive_positive_value_PL2_W}
\alias{predictive_positive_value_PL2_tW}
\alias{predictive_positive_value_PSel}
\alias{predictive_positive_value_PSel.5}
\alias{predictive_positive_value_PSel.e2}
\alias{predictive_positive_value_PSel.5.e2}
\alias{predictive_positive_value_PSel_W}
\alias{predictive_positive_value_robust}
\alias{predictive_positive_value_PB}
\alias{predictive_positive_value_PB_095_075}
\alias{predictive_positive_value_PB_075_075}
\alias{predictive_positive_value_PB_W}
\alias{F_score_C}
\alias{F_score_PL}
\alias{F_score_PL2}
\alias{F_score_PL2_W}
\alias{F_score_PL2_tW}
\alias{F_score_PSel}
\alias{F_score_PSel.5}
\alias{F_score_PSel.e2}
\alias{F_score_PSel.5.e2}
\alias{F_score_PSel_W}
\alias{F_score_robust}
\alias{F_score_PB}
\alias{F_score_PB_095_075}
\alias{F_score_PB_075_075}
\alias{F_score_PB_W}
\alias{nv_C}
\alias{nv_PL}
\alias{nv_PL2}
\alias{nv_PL2_W}
\alias{nv_PL2_tW}
\alias{nv_PSel}
\alias{nv_PSel.5}
\alias{nv_PSel.e2}
\alias{nv_PSel.5.e2}
\alias{nv_PSel_W}
\alias{nv_robust}
\alias{nv_PB}
\alias{nv_PB_095_075}
\alias{nv_PB_075_075}
\alias{nv_PB_W}
\title{Simulations for reverse-engineering}
\format{A numerical matrix frame with 100 rows and 200 variables or a numerical vector of length 100.}
\usage{
test.seq_C

test.seq_PL

test.seq_PL2

test.seq_PL2_W

test.seq_PL2_tW

test.seq_PSel

test.seq_PSel.5

test.seq_PSel.e2

test.seq_PSel.5.e2

test.seq_PSel_W

test.seq_robust

test.seq_PB

test.seq_PB_095_075

test.seq_PB_075_075

test.seq_PB_W

sensitivity_C

sensitivity_PL

sensitivity_PL2

sensitivity_PL2_W

sensitivity_PL2_tW

sensitivity_PSel

sensitivity_PSel.5

sensitivity_PSel.e2

sensitivity_PSel.5.e2

sensitivity_PSel_W

sensitivity_robust

sensitivity_PB

sensitivity_PB_095_075

sensitivity_PB_075_075

sensitivity_PB_W

predictive_positive_value_C

predictive_positive_value_PL

predictive_positive_value_PL2

predictive_positive_value_PL2_W

predictive_positive_value_PL2_tW

predictive_positive_value_PSel

predictive_positive_value_PSel.5

predictive_positive_value_PSel.e2

predictive_positive_value_PSel.5.e2

predictive_positive_value_PSel_W

predictive_positive_value_robust

predictive_positive_value_PB

predictive_positive_value_PB_095_075

predictive_positive_value_PB_075_075

predictive_positive_value_PB_W

F_score_C

F_score_PL

F_score_PL2

F_score_PL2_W

F_score_PL2_tW

F_score_PSel

F_score_PSel.5

F_score_PSel.e2

F_score_PSel.5.e2

F_score_PSel_W

F_score_robust

F_score_PB

F_score_PB_095_075

F_score_PB_075_075

F_score_PB_W

nv_C

nv_PL

nv_PL2

nv_PL2_W

nv_PL2_tW

nv_PSel

nv_PSel.5

nv_PSel.e2

nv_PSel.5.e2

nv_PSel_W

nv_robust

nv_PB

nv_PB_095_075

nv_PB_075_075

nv_PB_W
}
\description{
Result of fastboost analysis applied to biological network reverse engineering
}
\keyword{datasets}
