% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectBoost-package.R
\docType{package}
\name{SelectBoost}
\alias{SelectBoost}
\alias{SelectBoost-package}
\title{SelectBoost}
\description{
Motivation: With the growth of big data, variable selection has become one of the major
challenges in statistics. Although many methods have been proposed in the literature their
performance in terms of recall and precision are limited in a context where the number of
variables by far exceeds the number of observations or in a high correlated setting.
Results: This package implements a new general algorithm which improves the precision of any
existing variable selection method. This algorithm is based on highly intensive simulations and #' takes into account the correlation structure of the data. Our algorithm can either produce a
confidence index for variable selection or it can be used in an experimental design planning
perspective.
}
\references{
\emph{selectBoost: a general algorithm to enhance the performance of variable selection methods in correlated datasets}, Ismaïl Aouadi, Nicolas Jung, Raphael Carapito, Laurent Vallat, Seiamak Bahram, Myriam Maumy-Bertrand, Frédéric Bertrand, \url{https://arxiv.org/abs/1810.01670}
}
