% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoboost.R
\name{summary.selectboost}
\alias{summary.selectboost}
\title{Summarize a selectboost analysis}
\usage{
\method{summary}{selectboost}(object, crit.func = mean,
  crit.int = "mean", custom.values.lim = NULL, index.lim = NULL,
  alpha.conf.level = 0.99, force.dec = TRUE, ...)
}
\arguments{
\item{object}{Numerical matrix. Result of selectboost (autoboost, fastboost, ...).

selectboost_result,crit.func=mean,crit.int="mean",custom.values.lim=NULL,index.lim=NULL,alpha.conf.level=0.99,force.dec=TRUE,...}

\item{...}{. Passed to the plotting functions.}

\item{verbose}{Boolean.
Defaults to \code{FALSE}.}

\item{prop.level}{Numeric value. Used to compute the proportion of selection is
greater than prop.level. Defaults to \code{.95}.}

\item{conf.int.level}{Numeric value. Confidence level for confidence intervals on estimated
proportions of selection. Defaults to \code{.95}.}

\item{conf.threshold}{Numeric value. Used to compute the number of steps (c0) for which
the proportion of selection remains greater than conf.threshold. Defaults to \code{.95}.}
}
\value{
An invisible list.
}
\description{
Summarize a selectboost analysis.
}
\details{
\code{summary.selectboost} returns an invisible list and creates four graphics.
Two plots the proportion of selection with respect to c0 (by step or according to real scale).
On the third graph, no bar means a proportion of selection less than prop.level.
Confidence intervals are computed at the conf.int.level level.
Barplot of the confidence index (1-min(c0, such that proportion|c0>conf.threshold)).
}
\examples{
set.seed(314)
xran=matrix(rnorm(75),15,5)
ybin=sample(0:1,15,replace=TRUE)
yran=rnorm(15)
layout(matrix(1:4,2,2))

data(autoboost.res.x)
summary(autoboost.res.x)
summary(autoboost.res.x, force.dec=FALSE)

data(autoboost.res.x.adapt)
summary(autoboost.res.x.adapt)

data(autoboost.res.x2)
summary(autoboost.res.x2)
summary(autoboost.res.x2, force.dec=FALSE)

data(autoboost.res.x2.adapt)
summary(autoboost.res.x2.adapt)

data(fastboost.res.x)
summary(fastboost.res.x)
summary(fastboost.res.x, force.dec=FALSE)

data(fastboost.res.x.adapt)
summary(fastboost.res.x.adapt)

data(fastboost.res.x2)
summary(fastboost.res.x2)
summary(fastboost.res.x2, force.dec=FALSE)

data(fastboost.res.x2.adapt)
summary(fastboost.res.x2.adapt)
}
\references{
\emph{selectBoost: a general algorithm to enhance the performance of variable selection methods in correlated datasets}, Ismaïl Aouadi, Nicolas Jung, Raphael Carapito, Laurent Vallat, Seiamak Bahram, Myriam Maumy-Bertrand, Frédéric Bertrand, \url{https://arxiv.org/abs/1810.01670}
}
\seealso{
\code{\link{fastboost}}, \code{\link{autoboost}}

Other Selectboost analyse functions: \code{\link{plot.selectboost}}
}
\author{
Frederic Bertrand, \email{frederic.bertrand@math.unistra.fr}
}
\concept{Selectboost analyse functions}
