% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Cascade_example}
\alias{Cascade_example}
\alias{M}
\alias{Net}
\alias{Net_inf_C}
\title{Simulated Cascade network and inference}
\format{Three objects :
\describe{
  \item{M}{Simulated microarray}
  \item{Net}{Simulated network}
  \item{Net_inf_C}{Inferred network}
}}
\usage{
M

Net

Net_inf_C
}
\description{
Result for the reverse engineering of a simulated Cascade network
}
\keyword{datasets}
